+++ src/main/java/net/silentchaos512/gear/api/traits/ITraitInstance.java
@@ -9,6 +9,8 @@ import net.minecraft.util.text.StringTextComponent;
 import net.silentchaos512.gear.api.item.GearType;
 import net.silentchaos512.gear.api.part.PartDataList;
 import net.silentchaos512.gear.api.part.PartType;
+import net.silentchaos512.gear.api.util.IGearComponentInstance;
+import net.silentchaos512.gear.api.util.PartGearKey;
 import net.silentchaos512.gear.gear.material.MaterialInstance;
 import net.silentchaos512.gear.gear.trait.TraitSerializers;
 import net.silentchaos512.gear.util.TextUtil;
@@ -27,14 +29,21 @@ public interface ITraitInstance {
 
     Collection<ITraitCondition> getConditions();
 
+    @Deprecated
     default boolean conditionsMatch(GearType gearType, PartDataList parts, ItemStack gear) {
         ITrait trait = getTrait();
         return trait == null || getConditions().stream().allMatch(c -> c.matches(gear, gearType, parts, trait));
     }
 
+    @Deprecated
     default boolean conditionsMatch(List<MaterialInstance> materials, GearType gearType, PartType partType, ItemStack gear) {
         ITrait trait = getTrait();
         return trait == null || getConditions().stream().allMatch(c -> c.matches(gear, gearType, partType, materials, trait));
+    }
+
+    default boolean conditionsMatch(PartGearKey key, ItemStack gear, List<IGearComponentInstance<?>> components) {
+        ITrait trait = getTrait();
+        return trait == null || getConditions().stream().allMatch(c -> c.matches(trait, key, gear, components));
     }
 
     default JsonObject serialize() {