+++ src/main/java/net/silentchaos512/gear/gear/trait/condition/MaterialCountTraitCondition.java
@@ -14,6 +14,8 @@ import net.silentchaos512.gear.api.traits.ITrait;
 import net.silentchaos512.gear.api.traits.ITraitCondition;
 import net.silentchaos512.gear.api.traits.ITraitConditionSerializer;
 import net.silentchaos512.gear.api.traits.TraitInstance;
+import net.silentchaos512.gear.api.util.IGearComponentInstance;
+import net.silentchaos512.gear.api.util.PartGearKey;
 import net.silentchaos512.gear.gear.material.MaterialInstance;
 import net.silentchaos512.gear.util.TextUtil;
 
@@ -50,6 +52,20 @@ public class MaterialCountTraitCondition implements ITraitCondition {
         int count = 0;
         for (MaterialInstance mat : materials) {
             for (TraitInstance inst : mat.getTraits(partType, gearType, gear)) {
+                if (inst.getTrait() == trait) {
+                    count++;
+                    break;
+                }
+            }
+        }
+        return count >= this.requiredCount;
+    }
+
+    @Override
+    public boolean matches(ITrait trait, PartGearKey key, ItemStack gear, List<IGearComponentInstance<?>> components) {
+        int count = 0;
+        for (IGearComponentInstance<?> comp : components) {
+            for (TraitInstance inst : comp.getTraits(key, gear)) {
                 if (inst.getTrait() == trait) {
                     count++;
                     break;